src/requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.1
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.23.0
saucenao-api==2.4.0
six==1.16.0
typing_extensions==4.9.0
urllib3==1.25.11
Werkzeug==3.0.1
yarl==1.9.4
